backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
sqlalchemy
pydantic
gunicorn